[fix] loss of precision caused by cast type conversion

+++ utils/funcexp/func_time_to_sec.cpp
@@ -221,7 +221,7 @@ IDB_Decimal Func_time_to_sec::getDecimalVal(rowgroup::Row& row, FunctionParm& pa
   {
     if (scaleDiff > 0)
     {
-      value = (int64_t)round(value / IDB_pow[scaleDiff]);
+      value = (int64_t)round((double)value / IDB_pow[scaleDiff]);
     }
     else if (scaleDiff < 0)
     {
